Quick Animation Fix
TimeUnit.SECONDS.toNanos(1 / 30); was an incorrect call as it always returns 0, the correct call was TimeUnit.SECONDS.toNanos(1L) / 30; which returns 33,333,333 which is much closer to the real value of 1/30th of a second expressed as nano seconds.

This fixes the battle scene and now animation happens as expected, battle scenes are much slower as a side effect but it's more clear when an NPC is attacking or when they are finished attacking.