Fixes for JavaFXExample program
Run_Minimal.java is the minimal run environment with the TimeUnit fix applied.

Run.java and Light.java would run together and show how to draw on the canvas object using the GraphicsContext, as well as the basic principle of moving sprites around via an Objects own x and y variables.