v0.2.1-alpha mapMeta Rework
now the :FIRE and :IMPASSABLE meta tags will keep track of the tilesets in the map, so we can have multiple tilesets in one map and the meta tags will properly be assigned per tileset. Might rework the meta FIRE and IMPASSABLE to be included in the Constructor of the TileSet object and move that .meta data out. That way different maps using the same TileSet wouldn't have to re assign all the tags.