v0.2.1a - Different Resolution Settings
Almost have it working, just need to get it to not render NPCs that aren't in the visible game area. Render.java I have the map correctly rendering, the camera centers on the player entity. all I need to do is work out how to make it not render AI that aren't on screen.