v0.2.2a - JLink Image JRT FileSystem
By changing the way the we handle paths we have effectively allowed the program to detect if it is in a JLink image, and to use it's internal JRT filesystem.

The .zip included with this next release does not require the doubling of the /GameData/ folder being placed into /bin, now everything is done within the JLink images JRT filesystem. Might make a /mod folder that would allow people to use an external file system for making maps and such.