v0.2.3a
Added more potential states.
Added functioning EXIT_TO_MAIN screen and SETTINGS screens.
This update is a good example of how to add States, Screens, and Functionality to those.

Everything appears to be working for now. This update will also generate a /bin/Mod/ folder based on the result of Paths.get("") might need to tweak this later incase that doesn't point to /bin for this custom JRE for users who manipulate their class paths.