v0.2.3a - Map Icons & JRT
Now when you create a new map and save it from the JLink image it will generate the .dat and .meta files into /Mod/Maps. For now it only accesses the internal JRT tilesets, soon I'll make a system to tell if a tileset is internal or stored in the Mod/Art folder.

You can also generate Icons now from the JLink image and the icons will be stored in Mod/Art. The system can determine if the current map is being loaded from JRT or not, if it is it will search for Icons in the Mod/Art folder.

Have the beginning shell of an item system starting to creep up. Initally I'll probably just make an HP pickup item, and a Gold pickup item. MUCH later I will think about equipment such as weapons/armor. the HP and Gold items will motivate me to start thinking about post battle game flow, such as showing a results screen,  and perhaps adding an 'inbetween levels shop screen' type of system.