v0.2.3a - Mod Maps
Now if a /Mod/Maps .dat file exists and it has the tileset src paths marked with MOD.{tilsetName} it will know that it should be looking in the /Mod/Art path rather than the internal JRT. I left alot of debug output for this part in console so you can see the madness when you use it. I'll upload a release .zip with a pre made /Mod/Maps/Random .dat and .meta files as well as an example /Mod/Art _Icon as well has MOD.Rivery and MOD.Dunegoney tilsets to show how this system works.