v0.2.3a-Item System Beginnings
We can now load map 2 and it will render a Gold item onto the map at the location defined in the maps .meta file. We can reference the GameData\Items\0_baseItems.dat which will give us the needed base information on the Item.

This is in no way the final form of the system, I'm just suprised I got it working so easily. Probably introduces plenty of issues for JLink so I won't be publishing a release for this update, once I've got it all functional I'll update the releases.

Likely I will move things around in the Art folder, and work on how TileSets are created for different types of things. hence the inclusion of a Art\Item folder.

Map two is a little different then the empty grassland it was, It also shows off how we can use multiple tilesets in a single map.